requirements.txt
Flask~=3.0
feedparser~=6.0
requests~=2.32
python-dotenv~=1.0


